rename main module dir: resources -> core
