+++ core/loaders.py
