+++ grpc_protobuf/model_manager_grpc_pb2_grpc.py
@@ -4,7 +4,7 @@ import grpc
 import warnings
 
 from google.protobuf import empty_pb2 as google_dot_protobuf_dot_empty__pb2
-from grpc_protobuf import model_manager_grpc_pb2 as model__manager__grpc__pb2
+from . import model_manager_grpc_pb2 as model__manager__grpc__pb2
 
 GRPC_GENERATED_VERSION = '1.67.1'
 GRPC_VERSION = grpc.__version__
